ai-app/requirements.txt
fastapi
uvicorn
pillow
transformers
torch
easyocr
numpy
